Prevent browser from opening on ctrl+c on windows (#1120)
## Description

Exiting `airplane dev` currently causes the browser to open in Windows. This is because the interrupt signal is getting captured by `fmt.Scanln()`. Furthermore, we always advance past `fmt.Scanln()` _before_ the stop channel receives a notification of an interrupt signal. This PR adds an artificial timeout that allows the stop channel to get notified of the interrupt signal before continuing on to open the browser.

cc @yolken-airplane 

Resolves ASM-20

## Test plan
Verified that the browser-opening issue no longer occurs on windows.