Only prompt on required tsconfig changes (#1129)
## Description

Makes it so that the user is prompted only to modify their existing tsconfig only if they don't have certain required fields. This helps reduce noise for teams that don't want their tsconfig to be updated, and in general makes us less opinionated about what fields need to be inside the tsconfig. For now, the only configs that we require (for views) are "jsx" and "lib". cc @zhan-airplane 

If no tsconfig.json exists, the logic to write the default one remains the same.

Note that if a user has an invalid value in their tsconfig (let's say `"jsx": "blah"`), this doesn't fix it for the user as we're only checking that required properties _exist_.

I removed the logic to prefer the default TS config vs the existing TS config since we now always prefer the existing TS config no matter what.

I also added a test since this was previously untested. 

Resolves ASM-532

## Test plan
I developed + deployed a TS task, view, and template with an empty tsconfig and verified that everything worked. I verified that I'm only prompted to make a tsconfig update when there's a missing required field.