Add updatedAt to Views APIs (#1127)
## Description

This PR updates the CLI to start tracking when Views were last discovered. This allows us to return an `updatedAt` field for Views just like we already did for tasks. This is necessary for the Studio config panel which relies on the `updatedAt` field to determine if an update from the API should be applied to the local config panel state.

This PR also does a few renames: (1) removes a reference to App in AppConditions and (2) renames TasksConfigs to LocalTasks (to clarify it does not include remote tasks and because I changed the type to attach UpdatedAt).

This also changes how `UpdatedAt` is calculated for tasks. Previously, it was computed every time `GetTaskErrors` was recomputed, which would happen after every rediscovery (which clears the EntityCondition cache). This makes it harder to reason about what the timestamp represents (since it depends on when GetTaskErrors is next called). We weren't using `EntityCondition.RefreshedAt` for anything else, so I removed it in favor of `TaskState.UpdatedAt`.

## Test plan

Opened a task and view in studio and confirmed both had `updatedAt` set. Confirmed I could edit tasks and views from the config panel (which relies on `updatedAt`).