Move packages into `pkg/cli` (#1135)
## Description

This PR renames a handful of Go packages to prepare for merging with airport. In a few cases, the renames help avoid conflicts (e.g. both have a `versions` package). In other cases, it's to clarify the intention of each package (e.g. I moved a majority of CLI-specific packages into `pkg/cli/...`).

There are a few other minor changes that align the structure of this repo with airport.

Folks will need to update their aliases to point at `cmd/cli` instead of `cmd/airplane`. I'll post an update in `#eng` once this lands.

## Test plan

Tests should pass. Tested studio locally.